+++ Engine/porting/Win32/Platform/CWin32HttpStream.cpp
@@ -206,8 +206,8 @@ CWin32HttpStream::getWriteStream()
 	return NULL;
 }
 
-int
-CWin32HttpStream::readU16arr(u16 *pBufferU16, int items)
+size_t
+CWin32HttpStream::readU16arr(u16 *pBufferU16, size_t items)
 {
 	size_t len = sizeof(u16) * items;
 	int cnt = items;
@@ -222,8 +222,8 @@ CWin32HttpStream::readU16arr(u16 *pBufferU16, int items)
     return cnt;
 }
 
-int
-CWin32HttpStream::readU32arr(u32 *pBufferU32, int items)
+size_t
+CWin32HttpStream::readU32arr(u32 *pBufferU32, size_t items)
 {
 	size_t len = sizeof(u32) * items;
 	int cnt = items;
