+++ Engine/prebuilt/OSX/PlaygroundOSS.framework/Headers/FileSystem.h
@@ -44,8 +44,8 @@ public:
 	virtual		u16		readU16		()	= 0;	// Use cast to support s16
 	virtual		u32		readU32		()	= 0;	// Use cast to support s32
     
-    virtual     int     readU16arr  (u16 * pBufferU16, int items) = 0;
-    virtual     int     readU32arr  (u32 * pBufferU32, int items) = 0;
+    virtual     size_t  readU16arr  (u16 * pBufferU16, size_t items) = 0;
+    virtual     size_t  readU32arr  (u32 * pBufferU32, size_t items) = 0;
     
 	virtual		float	readFloat	()	= 0;	//
 	virtual		bool	readBlock	(void* buffer, u32 byteSize)	= 0;
